+++ orthoseg/util/vector_util.py
@@ -111,7 +111,7 @@ def simplify_topo_orthoseg(
     geoseries_copy = geoseries.copy()  # type: ignore
 
     # Set empty geometries to None, if no geometries are left, return
-    empty_idxs = pygeos.is_empty(geoseries_copy.array.data).nonzero()
+    empty_idxs = pygeos.is_empty(geoseries_copy.array.data).nonzero()  # type: ignore
     if len(empty_idxs) > 0:
         geoseries_copy.iloc[empty_idxs] = None
     if len(geoseries_copy[geoseries_copy != np.array(None)]) == 0: